refactor(core): extract interval selection controller
